Addition of Closed Lab tests with a lab that is closed always.
addition of _labTimezone function.
Creates a Date element with the users current location to be used in case the request fails or no result is returned
Set time with offset to make time to be the current time on the lab location. rawOffset is negative and contains the offset in seconds.